[core] Switched from CashdeskSessionItem to ItemMovement model
The model works the same way, but permits an item to be added (or
removed) to a session several times. This covers a workflow involving
resupplying a session in a transparent, accountable way.
It's a timestamped-ish model to help account for item movements, but the
timestamp is not protected, enforced or a security feature in any way.